Move individual and objective into problems
